an ElmHttpIdType constructor for ElmDatatype
this is a bit of a hack, but the idea is to allow id's which are
wrapped inside of a container to be passed as url values.

to use this, override the default interface of ElmType like so:

```haskell
instance ElmType Api.WorkbookId where
    toElmType wbId = ElmHttpIdType
                        (pack "WorkbookId")
                        (RecordConstructor (pack "WorkbookId") $
                         ElmField (pack "id") $ ElmPrimitiveRef EInt )
                        (pack "id")
```

I'm doing a lot of manual work the GenericElm* does for me, but I don't
understand reflection in haskell well enough to do it properly.  In the
above example, I am saying that the "id" field of a "Api.WorkbookId"
is what goes into a url parser in servant-elm.

If the ID is a string, one should set that as an option in servant-elm.